Add migrations section to sidebar with list component sub nav items
