+++ src/components/PlansList.tsx
@@ -1,26 +1,23 @@
 import * as React from "react";
-import { QueryClient, QueryClientProvider } from "react-query";
-import { Route, Switch } from "react-router-dom";
-import ExtensionConsumerChild from "./ExtensionConsumerChild";
-import LoginHandlerComponent from "./LoginHandlerComponent";
-import PrivateRoute from "./PrivateRoute";
+import { usePlansQuery } from "./plans";
 
 const PlansList: React.FC = () => {
-  const queryClient = new QueryClient();
-  // TODO: Handle Auth when the shared token is made available from the console.
-  // https://github.com/openshift/api/pull/1020
+  const plansQuery = usePlansQuery();
+  console.log("plansQuery", plansQuery);
+  // const [data, loaded] = useK8sWatchResource<K8sResourceCommon[]>({
+  //   kind: 'Pod',
+  //   isList: true,
+  // });
   return (
-    <QueryClientProvider client={queryClient}>
-      <Switch>
-        <Route path="/handle-login" component={LoginHandlerComponent} />
-        <PrivateRoute
-          path="/"
-          // isLoggedIn={!!auth.user}
-          isLoggedIn={true}
-          component={ExtensionConsumerChild}
-        />
-      </Switch>
-    </QueryClientProvider>
+    <div>
+      Plans list:
+      {plansQuery?.data?.items.map((d) => (
+        <div key={d.metadata.name}>{d.metadata.name}</div>
+      ))}
+      {/* <div>The WebSocket is currently {connectionStatus}</div>
+      {lastJsonMessage ? <span>Last message: {lastJsonMessage?.data?.kind}</span> : null}
+      <ul></ul> */}
+    </div>
   );
 };
